modules/job-manager: move cancel to raise.c
Rename cancel.[ch] to raise.[ch].

This change is purely cosmetic.
